requirements.txt
atlassian-python-api==3.19.0
bitbucket-python==0.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
deepdiff==5.7.0
Deprecated==1.2.13
future==0.18.2
idna==3.3
oauthlib==2.1.0
ordered-set==4.0.2
pybitbucket==0.12.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
python-dotenv==0.19.2
python-gitlab==3.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
simplejson==3.17.6
six==1.16.0
uritemplate==0.6
urllib3==1.26.8
voluptuous==0.12.2
wrapt==1.13.3
